+++ packages/std/src/results/migrate.rs
@@ -3,7 +3,6 @@ use serde::{Deserialize, Serialize};
 use std::fmt;
 
 use crate::binary::Binary;
-use crate::errors::StdError;
 use crate::types::Empty;
 
 use super::attribute::Attribute;
@@ -32,10 +31,3 @@ where
         }
     }
 }
-
-#[deprecated(
-    since = "0.12.1",
-    note = "MigrateResult is deprecated because it uses StdError, which should be replaced with custom errors in CosmWasm 0.11+. \
-            Replace this with Result<MigrateResponse, StdError> or Result<MigrateResponse<U>, StdError> and consider migrating to custom errors from there."
-)]
-pub type MigrateResult<U = Empty> = Result<MigrateResponse<U>, StdError>;